Test-Site/app/requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
attrs==19.2.0
azure-core==1.0.0
azure-storage-blob==12.0.0
azure-storage-file-datalake==12.0.0b5
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.2
Click==7.0
cryptography==2.8
cycler==0.10.0
dash==1.6.1
dash-bootstrap-components==0.7.1
dash-core-components==1.5.1
dash-daq==0.3.1
dash-html-components==1.0.2
dash-player==0.0.1
dash-renderer==1.2.1
dash-table==4.5.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.7
dnspython==1.16.0
entrypoints==0.3
Flask==1.1.1
Flask-Caching==1.7.2
Flask-Compress==1.4.0
future==0.18.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
gunicorn==19.7.1
h5py==2.10.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
moviepy==1.0.1
msrest==0.6.10
mysql-connector-python==8.0.18
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
plotly==4.3.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests-oauthlib==1.3.0
retrying==1.3.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
utils==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
